refactor(mcp): move superset_core MCP module from mcp to api/mcp
