+++ superset/mcp_service/chart/tool/generate_chart.py
@@ -23,7 +23,7 @@ import time
 from urllib.parse import parse_qs, urlparse
 
 from fastmcp import Context
-from superset_core.mcp import tool
+from superset_core.api.mcp import tool
 
 from superset.commands.exceptions import CommandException
 from superset.extensions import event_logger
