+++ superset/mcp_service/dashboard/tool/add_chart_to_existing_dashboard.py
@@ -25,7 +25,7 @@ import logging
 from typing import Any, Dict
 
 from fastmcp import Context
-from superset_core.mcp import tool
+from superset_core.api.mcp import tool
 
 from superset.extensions import event_logger
 from superset.mcp_service.dashboard.constants import (
