+++ superset/mcp_service/dataset/tool/list_datasets.py
@@ -26,7 +26,7 @@ import logging
 from typing import TYPE_CHECKING
 
 from fastmcp import Context
-from superset_core.mcp import tool
+from superset_core.api.mcp import tool
 
 if TYPE_CHECKING:
     from superset.connectors.sqla.models import SqlaTable
